requirements.txt
alpaca-py>=0.38.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
python-dotenv>=1.0.0
dash>=2.18.2
scipy>=1.15.0
